+++ spoon-runner/src/main/java/com/squareup/spoon/DeviceTestResult.java
@@ -11,13 +11,15 @@ import java.util.concurrent.TimeUnit;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static java.util.Collections.unmodifiableList;
+import static com.squareup.spoon.SpoonLogger.logError;
 
+/** Represents the result of a single test method on a single device. */
 public final class DeviceTestResult {
   /** Separator between screenshot timestamp and tag. */
   public static final String SCREENSHOT_SEPARATOR = Spoon.NAME_SEPARATOR;
 
   public enum Status {
-    PASS, FAIL, ERROR
+    PASS, FAIL
   }
 
   private final Status status;
@@ -85,16 +87,10 @@ public final class DeviceTestResult {
 
     public Builder markTestAsFailed(String message) {
       checkNotNull(message);
-      checkArgument(status == Status.PASS, "Status was already marked as " + status);
+      if (status != Status.PASS) {
+        logError("Status was already marked as failed!");
+      }
       status = Status.FAIL;
-      exception = StackTrace.from(message);
-      return this;
-    }
-
-    public Builder markTestAsError(String message) {
-      checkNotNull(message);
-      checkArgument(status == Status.PASS, "Status was already marked as " + status);
-      status = Status.ERROR;
       exception = StackTrace.from(message);
       return this;
     }
@@ -114,7 +110,9 @@ public final class DeviceTestResult {
 
     public Builder endTest() {
       checkArgument(start != 0, "Start was not called.");
-      checkArgument(duration == -1, "End was already called.");
+      if (duration != -1) {
+        logError("Test was already marked as ended!");
+      }
       duration = TimeUnit.NANOSECONDS.toSeconds(System.nanoTime() - start);
       return this;
     }